requirements.txt
streamlit
psycopg2-binary
sqlalchemy
openai
python-dotenv
